pyproject.toml
[tool.poetry]
name = "dados-financeiros-b3"
version = "0.1.0"
description = ""
authors = ["Prog-LucasAlves <lucasalves_taz@hotmail.com>"]
readme = "README.md"
packages = [{include = "dados_financeiros_b3"}]

[tool.poetry.dependencies]
python = "^3.11"
backoff = "^2.2.1"
pandas = "^1.5.2"
requests = "^2.28.1"
beautifulsoup4 = "^4.11.1"
tqdm = "^4.64.1"
selenium = "^4.7.2"
python-dotenv = "^0.21.0"
psycopg2 = "^2.9.5"
yfinance = "^0.2.3"
quantstats = "^0.0.59"
streamlit = "^1.16.0"
plotly = "^5.11.0"
seaborn = "^0.12.2"
pychatgpt = "^0.4.3.1"
openpyxl = "^3.1.2"
xlsxwriter = "^3.0.9"
flask = "^2.3.2"


[tool.poetry.group.dev.dependencies]
ipykernel = "^6.19.4"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
